SSL cleanup
