+++ cpp/include/Ice/Ice.h
@@ -34,10 +34,10 @@
 #    include "Properties.h"
 #    include "ProxyFunctions.h"
 #    include "RegisterPlugins.h"
-#    include "SSL/Certificate.h"
 #    include "SSL/ClientAuthenticationOptions.h"
 #    include "SSL/ConnectionInfo.h"
 #    include "SSL/EndpointInfo.h"
+#    include "SSL/Exception.h"
 #    include "SSL/ServerAuthenticationOptions.h"
 #    include "ServantLocator.h"
 #    include "SlicedData.h"
