+++ cpp/src/Ice/SSL/DistinguishedName.h
@@ -2,78 +2,17 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 //
 
-#ifndef ICE_SSL_CERTIFICATE_H
-#define ICE_SSL_CERTIFICATE_H
+#ifndef ICE_SSL_DISTINGUISHED_NAME_H
+#define ICE_SSL_DISTINGUISHED_NAME_H
 
-#include "../Config.h"
-#include "../Exception.h"
-#include "ConnectionInfoF.h"
+#include "Ice/Config.h"
 
-#include <chrono>
-#include <cstdint>
-#include <functional>
 #include <list>
-#include <vector>
+#include <string>
+#include <utility>
 
 namespace Ice::SSL
 {
-    /**
-     * Thrown if the certificate cannot be read.
-     */
-    class ICE_API CertificateReadException : public Ice::Exception
-    {
-    public:
-        using Ice::Exception::Exception;
-
-        CertificateReadException(const char*, int, std::string) noexcept;
-
-        std::string ice_id() const override;
-
-        /** The reason for the exception. */
-        std::string reason;
-
-    private:
-        static const char* _name;
-    };
-
-    /**
-     * Thrown if the certificate cannot be encoded.
-     */
-    class ICE_API CertificateEncodingException : public Ice::Exception
-    {
-    public:
-        using Ice::Exception::Exception;
-
-        CertificateEncodingException(const char*, int, std::string) noexcept;
-
-        std::string ice_id() const override;
-
-        /** The reason for the exception. */
-        std::string reason;
-
-    private:
-        static const char* _name;
-    };
-
-    /**
-     * This exception is thrown if a distinguished name cannot be parsed.
-     */
-    class ICE_API ParseException : public Ice::Exception
-    {
-    public:
-        using Ice::Exception::Exception;
-
-        ParseException(const char*, int, std::string) noexcept;
-
-        std::string ice_id() const override;
-
-        /** The reason for the exception. */
-        std::string reason;
-
-    private:
-        static const char* _name;
-    };
-
     /**
      * This class represents a DistinguishedName, similar to the Java
      * type X500Principal and the .NET type X500DistinguishedName.
