+++ cpp/src/Ice/SSL/OpenSSLTransceiverI.h
@@ -8,9 +8,7 @@
 #include "../Network.h"
 #include "../StreamSocket.h"
 #include "../Transceiver.h"
-#include "../WSTransceiver.h"
 #include "Ice/Config.h"
-#include "Ice/SSL/Certificate.h"
 #include "Ice/SSL/ClientAuthenticationOptions.h"
 #include "Ice/SSL/ServerAuthenticationOptions.h"
 #include "OpenSSLEngineF.h"
