+++ cpp/src/Ice/SSL/SSLEngine.h
@@ -10,8 +10,8 @@
 #include "Ice/CommunicatorF.h"
 #include "Ice/Initialize.h"
 #include "Ice/InstanceF.h"
-#include "Ice/SSL/Certificate.h"
 #include "Ice/SSL/ClientAuthenticationOptions.h"
+#include "Ice/SSL/Exception.h"
 #include "Ice/SSL/ServerAuthenticationOptions.h"
 #include "SSLEngineF.h"
 #include "SSLInstanceF.h"
