+++ cpp/src/Ice/SSL/SecureTransportUtil.cpp
@@ -5,7 +5,7 @@
 #include "SecureTransportUtil.h"
 #include "../Base64.h"
 #include "Ice/LocalException.h"
-#include "Ice/SSL/Certificate.h"
+#include "Ice/SSL/Exception.h"
 #include "Ice/UniqueRef.h"
 #include "IceUtil/FileUtil.h"
 #include "IceUtil/StringUtil.h"
